Update and rename gcalculator.rb to gnome_calculator.rb
